src/com/twu/biblioteca/Book.java
package com.twu.biblioteca;

public class Book {

    private String author = "Unknown";
    private String title = "Unknown";
    private int year = 0;

    public String getTitle() {
        return title;
    }

    public void setTitle(String title) {
        this.title = title;
    }

    public void setAuthor(String author) {
        this.author = author;
    }

    public String getAuthor() {
        return this.author;
    }

    public int getYear() {
        return year;
    }

    public void setYear(int year) {
        this.year = year;
    }
}
